tokenizer: token generator and dump test tool
First pieces of the front end. We turn the source code into a token stream.
Process string quoting, constants into a useful form.

The only oddity here is that we pull all the names out of the stream and turn
them and number identical text names. There is no context associated wiht them
so symbol 14 may well be "i" in many places but not necessarily the same "i"
in C terms. It still makes the next stage much more efficient because we can
do the recursive descent parser entirely in terms of simple comparisons and
we can avoid keeping all the symbol text in memory as we only need it for the
code generator.