TODO: note more work to do on stack
In particular we want locals to work down the frame probably not start
at the bottom of the frame. It's a trade off. Putting them at the bottom of
the frame makes access via range limited stack referencing easier, and is
a big factor for some processors. Putting them at the top makes building
initializers nice if we teach the backend about evaluate and stack patterns.