backend: update to allow for labels as data objects
We need this so we can add in the changes needed for local static objects. These cannot use
the name because you can have two different static variables with local scope and the same
name.

Also move things to lval2 so we can fix the current mess with offset of labels.