cc: start enabling 'cconly'
Tell the backend when things are just for flag use so it can
instead construct simpler subtrees that just work the condition
flags.

This will need a fair bit more work to handle anything but simple
cases because subtrees of a || b && c etc need to be dealt with
as whole somewhere.