cc: big rework so we can optimize bools
Primarily we introduce the idea of a flag to indicate CC is needed. That
allows us to optimize bool in trees without losing the condition code
requirements wrongly. The rest of the work then uses this to optimize things
like if, do while and while where there are known constant cases.