z80: add framepointer mode and some small optimizations
We use frame pointers when we see > 4 bytes of locals and the user hasn't
used register to assign ix and iy to stuff. We don't have the space to do
the kind of analysis a modern compiler can so this is mostly a guess. It's
certainly going to be thrown by code with a lot of references to arguments