65c816: start fixing the basics
We now pass a few of the tests but a lot fail.

Fixing this also needed a fudge of the two ways we do a helper between X and A
as they ended up with the arguments different ways around. This needs revisiting
to generate better code. Probably we need to reverse the direct version, or
use the inverted helpers in one case.

Code generation is not yet particularly good so even if does work don't expect
anything epic to come out of the code generator.

Density is also a problem and may be trickier than I'd want to fix

No support yet for running data in bank 0 and using the real CPU stack for
C stack for faster denser code.