z8: a load more size optimization things
(and some speed)

The code is still a lot larger than we would like but it's hard
to see what we can do about that without a much smarter compiler
that won't fit into an 8bit machine.

ANSI pcc could do a much better job if taught for example as you could treat
most of the 128 odd bytes of register space as a giant working pool and take
the hit on interrupt and task switches.

At minimum we need to tune the code being built with "register" much more
and support the gcc like "register global" meaning zero page - for 680x,
6502 especially and Z8. (Maybe also 80196 when we get to that port).