65c816: add more -Os helpers, fix a lot of confusion about rep/sep
We had cases we set 8bit then transferred indexes around. Also confusion
when there was a pending switch to 16bit before fall through to a label.

This gets us closer (still the same fp error as without -Os), plus an error
in utol and shifts and a hang in qsort