super8: first pieces in place
The super8 has word register/register loads, slightly different ldei style
operators and some extremely useful register indexed and directly addressed
external memory operators.

Turn the basics on for now. There is a load more we can do with the bit ops
it the same way we do on Z80 (and for some compares maybe). That can follow.