6800: start on the 6800
This is the base of the 6800/3/HC11 and 6303. Start with the least functional
part but think about what is needed for the others in case we decide to
expand this to 680x instead of fixing up the initial sketch made during
the compiler design.

The output at this point is terrible, but that is fine. There is a lot of
low hanging fruit that as it is added will sort things out.