6800: more optimized subtrees
Because we do all our pointer referencing via X we get a free
addition in the reference. Therefore we rewrite assignment and dereference
to allow for an offset argument. We then fold deref(plus(?, constant)) and
similarly for eq.

Secondly fold up the case where this is all from a local and we can chain
values in X.

Finally spot the rather specific but common case of "*autovar = 0" as we
can use clr n,x in this case