target: add the bits needed to allow for word machines and structs
The problem we get is this

struct {
	int x;
	char y;
	int z;
} foo;

return foo.z on a word based machine usually needs to generate
load register with (_foo + 2) even though the struct offect is 4 bytes

but foo.y is a byte pointer so is _foo as a bytepointer (which is not
the same thing as _foo as a word pointer) + 2 (bytes)

On Nova for example we end up with

	jsr @__const1,0
	.word _foo + 2

and
	jsr @__const1,0
	.byteptr _foo + 2

We punt this lot to the target. Byte targets just generate an addition of
the base of the object with the offset as they did before.