nova: fix some more bugs
With these fixed we pass the base test set on the Nova. There will be plenty
more bugs left, and there is a lot of optimization work needed (eg register
value tracking, and not reloading without reason).

We are also still rewriting a bunch of stuff of the form

	lda 1,blah
	psha 1
	lda 1,blah
	popa 0

in the peepholes which probably should be coming out of gen_direct as 0/1
loads.