cc: change how .symtmp is handled a bit
The symtmp file is randomly named so unlike the other temp files doesn't end
up being cleaned up or replaced by later runs. It is however useful for
some debug cases. Instead of saving it with keep temp require the user specifies
the option at least twice to keep it. That means that the usual keep temp
behaviour cleans up nicely and is what developers are likely to want but the
.symtmp can still be saved when debugging crashes in a particular pass of
the compiler.