cc: go to 1024 sym nodes
This is annoying but it makes a big difference on what we can compile.

Need to look at the type handling and make sure all the matching func
prototypes are beign correctly merged, and some other size improvements
to compensate or we'll not be able to make it fit on boxes with poor code
density.