expression: correct flagging of varargs
In the case of cprintf(char *, ...) or similar being invoked with
cprint("foo") the argument walk would not actually hit the T_ELLIPSIS
and whilst the calling code correctly understood that it was ok if the
next argument was T_ELLIPSIS it did not then mark the call as varargs.

The 6800 backend cares about this case so broke as a result.

Closes: #139