tree: fix type handling problems with constify
We had several cases where would do

	n = tree(....)
	n->type = ...

However constify gets run before the assignment in tree and this caused
wrong optimizations and cast removals. Instead introduce a typed_tree
function and wrap it with tree. Convert the functions that use tree
and then mash the type to use typed_tree()

Reported-by: zu2