65c816: sort out float support
The 65C816 port does some strange things with argument stack handling in order
to stop it sucking totally. These are also used for internal helpers written in
C. That included float operations that returned ISBOOL (eg comparisons) but did
not contain the needed code to avoid trying to bool an ISBOOL of a float (which
isn't a good idea).

Rearrange the argstack helper code a bit so we can codegen the node to push and
thus find out if it's ISBOOL, and if so suppress the push and call.