expression: fix mishandling of vararg remaps
If you have a vararg argument then char goes to int and float
to double. However these then need rewriting into the target
actual types which was not happening. This caused a problem with

extern void x();

void y(float a)
{
	x(a);
}

as it tried to push a double even if the target doesn't have actual doubles