6800: fix mishandling of some some |= style ops
If we have a local variable as the right side and a local pointer as
the left we trashed X and wrote it back to the wrong place.

This is sufficient to get Fuzix booting into init although things then
go odd in init and that needs investigation to see if it's compiler
or kernel stuff.