6800/6809: Fix working value tracking bug with ++/--
x++ and x-- are odd in that the result is the old value unlike ++x or any +=
type operator. In the specific case where the value from an x++/x-- operation
was the working value and it was immediately reused then in some cases we would
mistakenly think we already had the right value loaded

This fixes the weird init misbehaviour when trying to build Fuzix.