tms7000: still a work in progress
Debug the accessors and tracking. We end up a bit differently to the Z8 as we have
decw but not incw and this messes up our memory handling a lot. Once it works we need
to look at a lot of places where we could fold/avoid additional add ops