8070: fix confusion about references
For can_make_ref purposes it matters if something is source or destination
to the caller so don't blindly use map_op.

This fixes the mishandling of

    *a++ = b = thing;

where it would wrongly eliminate b = thing on the grounds it could make
a reference to the LSTORE of b, whereas only an LREF would do.