6502: start adding bits for byteable and helpers
At the moment we fail on the byte sized shift tests because we don't clear
X and @tmp+1 but fail to do a byte sized != test. Need to look at what is
getting confused in the byteable logic here.

This patch tries to catch any byteable helpers in these cases and make them
byte sized but it's not the whole issue so needs further digging.